Changed where Tarkov Helper gets its data
Now using SP-Tarkov instead of JET for getting data files since they seems to update faster and more frequently just a drag and drop of the database folder from the server download and its good to go. Will update the wiki with a more in-depth reading on how to set it up once closer to release.